requirements.txt
PyQt5
transformers
torch
datasets
psutil
PyPDF2
onnx
safetensors
llama-cpp-python
accelerate
huggingface_hub
requests
pyinstaller
pytest
pdfplumber
python-docx
beautifulsoup4
pandas
openpyxl
pytesseract
pillow
arabic-reshaper
python-bidi
langdetect
ocrmypdf
easyocr
opencv-python-headless
pyspellchecker
camelot-py
tabula-py
